requirements.txt
absl-py==1.0.0
adal==1.2.7
antlr4-python3-runtime==4.13.1
appdirs==1.4.3
applicationinsights==0.11.10
argcomplete==3.1.6
astunparse==1.6.3
attrs==23.1.0
azure-appconfiguration==1.1.1
azure-batch==14.0.0
azure-cli==2.54.0
azure-cli-core==2.54.0
azure-cli-telemetry==1.1.0
azure-common==1.1.28
azure-core==1.29.5
azure-cosmos==3.2.0
azure-data-tables==12.4.0
azure-datalake-store==0.0.53
azure-graphrbac==0.60.0
azure-identity==1.15.0
azure-keyvault-administration==4.4.0b2
azure-keyvault-certificates==4.7.0
azure-keyvault-keys==4.9.0b3
azure-keyvault-secrets==4.7.0
azure-loganalytics==0.1.1
azure-mgmt-advisor==9.0.0
azure-mgmt-apimanagement==4.0.0
azure-mgmt-appconfiguration==3.0.0
azure-mgmt-appcontainers==2.0.0
azure-mgmt-applicationinsights==1.0.0
azure-mgmt-authorization==4.0.0
azure-mgmt-batch==17.0.0
azure-mgmt-batchai==7.0.0b1
azure-mgmt-billing==6.0.0
azure-mgmt-botservice==2.0.0
azure-mgmt-cdn==12.0.0
azure-mgmt-cognitiveservices==13.5.0
azure-mgmt-compute==30.3.0
azure-mgmt-containerinstance==10.1.0
azure-mgmt-containerregistry==10.1.0
azure-mgmt-containerservice==27.0.0
azure-mgmt-core==1.4.0
azure-mgmt-cosmosdb==9.3.0
azure-mgmt-databoxedge==1.0.0
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-datamigration==10.0.0
azure-mgmt-devtestlabs==4.0.0
azure-mgmt-dns==8.0.0
azure-mgmt-eventgrid==10.2.0b2
azure-mgmt-eventhub==10.1.0
azure-mgmt-extendedlocation==1.0.0b2
azure-mgmt-hdinsight==9.0.0
azure-mgmt-imagebuilder==1.2.0
azure-mgmt-iotcentral==10.0.0b2
azure-mgmt-iothub==3.0.0
azure-mgmt-iothubprovisioningservices==1.1.0
azure-mgmt-keyvault==10.3.0
azure-mgmt-kusto==0.3.0
azure-mgmt-loganalytics==13.0.0b4
azure-mgmt-managedservices==1.0.0
azure-mgmt-managementgroups==1.0.0
azure-mgmt-maps==2.0.0
azure-mgmt-marketplaceordering==1.1.0
azure-mgmt-media==9.0.0
azure-mgmt-monitor==5.0.1
azure-mgmt-msi==7.0.0
azure-mgmt-netapp==10.1.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-policyinsights==1.1.0b4
azure-mgmt-privatedns==1.0.0
azure-mgmt-rdbms==10.2.0b12
azure-mgmt-recoveryservices==2.5.0
azure-mgmt-recoveryservicesbackup==7.0.0
azure-mgmt-redhatopenshift==1.4.0
azure-mgmt-redis==14.1.0
azure-mgmt-resource==23.1.0b2
azure-mgmt-search==9.1.0
azure-mgmt-security==5.0.0
azure-mgmt-servicebus==8.2.0
azure-mgmt-servicefabric==1.0.0
azure-mgmt-servicefabricmanagedclusters==1.0.0
azure-mgmt-servicelinker==1.2.0b1
azure-mgmt-signalr==2.0.0b1
azure-mgmt-sql==4.0.0b12
azure-mgmt-sqlvirtualmachine==1.0.0b5
azure-mgmt-storage==21.1.0
azure-mgmt-synapse==2.1.0b5
azure-mgmt-trafficmanager==1.0.0
azure-mgmt-web==7.0.0
azure-multiapi-storage==1.2.0
azure-nspkg==3.0.2
azure-storage-blob==12.19.0
azure-storage-common==1.4.2
azure-synapse-accesscontrol==0.5.0
azure-synapse-artifacts==0.17.0
azure-synapse-managedprivateendpoints==0.4.0
azure-synapse-spark==0.2.0
bcrypt==4.1.1
cachetools==5.3.1
certifi==2019.11.28
cffi==1.16.0
chardet==3.0.4
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
cryptography==41.0.7
dask==2023.12.0
dbus-python==1.2.16
Deprecated==1.2.14
distlib==0.3.7
distributed==2023.12.0
distro==1.4.0
distro-info==0.23+ubuntu1.1
fabric==2.7.1
filelock==3.12.4
flatbuffers==23.1.21
fsspec==2023.10.0
gast==0.4.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.49.1
h5py==3.8.0
humanfriendly==10.0
idna==2.8
importlib-metadata==6.8.0
invoke==1.7.3
isodate==0.6.1
javaproperties==0.5.2
jax==0.3.25
jaxlib==0.3.25
Jinja2==3.1.2
jmespath==1.0.1
jsondiff==2.0.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
keras==2.13.1
knack==0.11.0
libclang==16.0.6
locket==1.0.0
Markdown==3.4.4
MarkupSafe==2.1.3
more-itertools==4.2.0
msal==1.25.0
msal-extensions==1.0.0
msgpack==1.0.7
msrest==0.7.1
msrestazure==0.6.4
numexpr==2.8.7
numpy==1.22.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.1.3
paramiko==3.3.1
partd==1.4.1
pathlib2==2.3.7.post1
pkginfo==1.9.6
platformdirs==3.10.0
portalocker==2.8.2
portpicker==1.4.0
protobuf==3.20.3
psutil==5.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycomposefile==0.0.30
pycparser==2.21
PyGithub==1.59.1
Pygments==2.17.2
PyGObject==3.36.0
PyJWT==2.8.0
PyNaCl==1.5.0
pyOpenSSL==23.3.0
PySocks==1.7.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.1
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rpds-py==0.13.2
rsa==4.9
scipy==1.7.3
scp==0.13.6
semver==2.13.0
six==1.16.0
sortedcontainers==2.4.0
ssh-import-id==5.10
sshtunnel==0.1.5
tabulate==0.9.0
tblib==3.0.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-rocm @ file:///tmp/tensorflow_pkg/tensorflow_rocm-2.13.0.570-cp39-cp39-linux_x86_64.whl#sha256=8492e103014f715f83899cf5d62011ec9916c13922a544f9a8ef542ce2cf254b
termcolor==2.1.1
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
unattended-upgrades==0.1
urllib3==1.25.8
virtualenv==20.24.5
websocket-client==1.3.3
Werkzeug==2.3.7
wrapt==1.14.1
xmltodict==0.13.0
zict==3.0.0
zipp==1.0.0
